Support Dataset size larger than 4 GiB Support arbitrary value of RANDOMX_DATASET_EXTRA_SIZE

+++ src/virtual_machine.hpp
@@ -60,7 +60,7 @@ protected:
 		randomx_cache* cachePtr = nullptr;
 		randomx_dataset* datasetPtr;
 	};
-	uint32_t datasetOffset;
+	uint64_t datasetOffset;
 };
 
 namespace randomx {
